A bit of code cleanup
