+++ Octokit/Helpers/WebHookConfigComparer.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+
+namespace Octokit
+{
+    public class WebHookConfigComparer : IEqualityComparer<KeyValuePair<string, string>>
+    {
+        public bool Equals(KeyValuePair<string, string> x, KeyValuePair<string, string> y)
+        {
+            return x.Key == y.Key;
+        }
+
+        public int GetHashCode(KeyValuePair<string, string> obj)
+        {
+            return obj.Key.GetHashCode();
+        }
+    }
+}
